Initial splitter code
