+++ bin/planet-stream.js
@@ -2,9 +2,11 @@
 
 // start planet-stream
 var argv = require('minimist')(process.argv.slice(2));
+
 var R = require('ramda');
 var planetstream = require('../')({
   verbose: argv.v,
+  limit: argv.limit,
   host: process.env.REDIS_PORT_6379_TCP_ADDR || process.env.REDIS_HOST || '127.0.0.1',
   port: process.env.REDIS_PORT_6379_TCP_PORT || process.env.REDIS_PORT || 6379
 });
