+++ lib/planetstream.js
@@ -2,6 +2,7 @@ var MetaStream = require('./streams/MetaStream.js');
 var DataStream = require('./streams/AugmentedDiffStream.js');
 var Redis = require('ioredis');
 var K = require('kefir');
+var splitter = require('./utils/splitter.js');
 
 function PlanetStream (opts) {
   if (!(this instanceof PlanetStream)) return new PlanetStream(opts);
@@ -85,7 +86,13 @@ function PlanetStream (opts) {
   })
   .map(JSON.stringify);
 
-  return joinedElements;
+  if (opts && opts.limit > 1) {
+    return joinedElements.flatMap(function (changeset) {
+      return splitter(changeset, opts.limit);
+    });
+  } else {
+    return joinedElements;
+  }
 }
 
 module.exports = PlanetStream;
