+++ lib/streams/AugmentedDiffStream.js
@@ -2,7 +2,7 @@ var K = require('kefir');
 var request = require('request-promise');
 var R = require('ramda');
 var AugmentedDiffParser = require('../parsers/AugmentedDiffParser.js');
-var log = require('../logger.js');
+var log = require('../utils/logger.js');
 
 /* Function that takes an options object and returns
  * an augmented diff stream with an associated 'current state'
