+++ lib/streams/MetaStream.js
@@ -3,7 +3,7 @@ var request = require('request-promise');
 var R = require('ramda');
 var zlib = require('zlib');
 var MetaParser = require('../parsers/MetaParser.js');
-var log = require('../logger.js');
+var log = require('../utils/logger.js');
 
 /* Takes a number representing the state and
  * outputs in the format xxx/xxx/xxx with zero
