+++ lib/utils/kinesis.js
@@ -2,7 +2,7 @@
 
 var Promise = require('bluebird');
 var shortid = require('shortid');
- 
+
  var AWS = require('aws-sdk');
 AWS.config.region = process.env.AWS_REGION;
 kin = new AWS.Kinesis()
@@ -18,7 +18,7 @@ function ListStreams() {
 function DescribeStream(name) {
     return pkinesis.describeStreamAsync({StreamName: name}).then(function(result) {
         return result.StreamDescription;
-    });    
+    });
 }
 
 function StreamStatus(name) {
